mini-twitter-api/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.6
Pillow==10.0.0
django-cors-headers==4.2.0
drf-yasg==1.21.7
celery==5.3.1
redis==4.6.0
django-redis==5.3.0
pytest==7.4.0
pytest-django==4.5.2
gunicorn==21.2.0
dj-database-url==2.1.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0